+++ app/build.gradle
@@ -72,8 +72,6 @@ android {
         versionName "0.1.4"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
 
-        buildConfigField "String", "TMDB_API_KEY", "\"" + propOrDef("TIVI_TMDB_API_KEY", "") + "\""
-
         playAccountConfig = playAccountConfigs.defaultAccountConfig
 
         manifestPlaceholders += ['appAuthRedirectScheme': 'me.banes.chris.tivi']
@@ -129,8 +127,7 @@ android {
 dependencies {
     implementation project(":data")
     implementation project(":trakt")
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre8:${versions.kotlin}"
+    implementation project(":tmdb")
 
     implementation "android.arch.lifecycle:runtime:1.0.3"
     implementation "android.arch.lifecycle:extensions:1.0.0-rc1"
@@ -146,8 +143,6 @@ dependencies {
     implementation "com.android.support.constraint:constraint-layout:${versions.constraint_layout}"
 
     implementation 'com.gabrielittner.threetenbp:lazythreetenbp:0.2.0'
-
-    implementation "com.uwetrottmann.tmdb2:tmdb-java:1.8.1"
 
     implementation "com.google.dagger:dagger-android:${versions.dagger}"
     implementation "com.google.dagger:dagger-android-support:${versions.dagger}"
