+++ app/src/main/java/me/banes/chris/tivi/home/HomeActivityViewModel.kt
@@ -17,10 +17,10 @@
 package me.banes.chris.tivi.home
 
 import android.arch.lifecycle.LiveData
-import me.banes.chris.tivi.calls.TmdbShowFetcher
 import me.banes.chris.tivi.data.daos.TiviShowDao
 import me.banes.chris.tivi.data.entities.TiviShow
 import me.banes.chris.tivi.extensions.plusAssign
+import me.banes.chris.tivi.tmdb.TmdbShowFetcher
 import me.banes.chris.tivi.trakt.TraktManager
 import me.banes.chris.tivi.util.AppRxSchedulers
 import me.banes.chris.tivi.util.RxAwareViewModel
