+++ tmdb/src/main/java/me/banes/chris/tivi/tmdb/TmdbModule.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2017 Google, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.banes.chris.tivi.tmdb
+
+import com.uwetrottmann.tmdb2.Tmdb
+import dagger.Module
+import dagger.Provides
+import okhttp3.Cache
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import java.io.File
+import javax.inject.Named
+import javax.inject.Singleton
+
+@Module
+class TmdbModule {
+
+    @Singleton
+    @Provides
+    fun provideTmdb(@Named("cache") cacheDir: File, interceptor: HttpLoggingInterceptor): Tmdb {
+        return object : Tmdb(BuildConfig.TMDB_API_KEY) {
+            override fun setOkHttpClientDefaults(builder: OkHttpClient.Builder) {
+                super.setOkHttpClientDefaults(builder)
+                builder.apply {
+                    addInterceptor(interceptor)
+                    cache(Cache(File(cacheDir, "tmdb_cache"), 10 * 1024 * 1024))
+                }
+            }
+        }
+    }
+
+}
